feat: Make logos adjustable
in addition to the main logo, other used images and icons can be
adjusted in the config.
Having require() in the config file was throwing some errors in Webpack 5.
Thus, we switch to an approach to resolve the image source path
during runtime using the config.

Add customization for:
- favicon
- brandLogo
- countryIcon (sidebar poi search)

Changes:
- add function imgSrc to utils.js for resolving the image source path
- rename logo@2x.png to logo.png
- add favicon id to dev.html
- small fixes