fix(route-information-popup): popup text splits mid-word
Previously floating features were used for the layout of the popups of route information.
As the size of the leaflet popup is based on the size of the div, when the icon for the
profile is floated, which moves the text but doesn't change the size of the underlying div.
This meant that the text got wrapped into new lines.

Now a flex layout is used for consistent positioning together with computing the
min-width from the content, to ensure the parent div resizes as expected and word breaks
are avoided.

- reduce to necessary css classes
- inline tooltip variable
- remove overflow-x/y from .leaflet-popup-content override to avoid scrollbars

Co-authored-by: Jakob Schnell <koebi@ezelo.de>