fix: hot-fix for compatibility issues with disaster-ors plugin
as the disaster ors plugin adds polygons with holes to the map
and the avoid areas currently collect all existing polygons,
area generation fails as the polygon with hole can't be transformed
to GeoJSON properly.
This simple check will ensure that the polygons with holes will
be skipped.

Efforts to put the avoid areas in their own layer group were
fruitless so far, as the L.Control.Draw requires them to be
present directly in the map overlay pane.