fix: remove loading of jobs from places
When moving from the places mode to the optimization tab, the first click anywhere on the map created a job from the place saved in the previous URL.
If no place was actively selected, this was the map centre.
To the user this may seem as though a job was created out of nowhere and appears as buggy behaviour, which is why it is removed for now.