fix: sidebar hiding on right-click
- hide sidebar on click only for low resolution as it doesn't make sense on larger screens
- adjust showMapViewClickPopups to split between 'med & high' vs 'low' instead of 'high' vs 'low & med'

if a right-click was done within 2 seconds after adding the first point, show click popups would still be false,
due to this dubious async logic that is now removed.

The logic somehow prevented to show the "What's this" modal, which without it came up when clicking an alternative
route.
This is prevented by stopping the propagation to the map base layer in alternativeRouteIndexSelected.